Move WritableFiles to irio and remove io artifact entierly
