+++ build.sbt
@@ -4,8 +4,6 @@ val scalajs = Build.root
 val ir = Build.irProject
 val irJS = Build.irProjectJS
 val compiler = Build.compiler
-val io = Build.io
-val ioJS = Build.ioJS
 val logging = Build.logging
 val loggingJS = Build.loggingJS
 val linker = Build.linker
