+++ linker/jvm/src/main/scala/org/scalajs/linker/backend/closure/ClosureLinkerBackend.scala
@@ -26,13 +26,13 @@ import com.google.javascript.jscomp.{
   _
 }
 
-import org.scalajs.io._
 import org.scalajs.logging.Logger
 
 import org.scalajs.linker._
-import org.scalajs.linker.standard._
 import org.scalajs.linker.backend._
 import org.scalajs.linker.backend.emitter.Emitter
+import org.scalajs.linker.irio._
+import org.scalajs.linker.standard._
 
 /** The Closure backend of the Scala.js linker.
  *
