+++ linker/shared/src/main/scala/org/scalajs/linker/standard/StandardLinkerImpl.scala
@@ -17,7 +17,6 @@ import scala.concurrent._
 import java.util.concurrent.atomic.AtomicBoolean
 
 import org.scalajs.logging.Logger
-import org.scalajs.io._
 
 import org.scalajs.linker._
 import org.scalajs.linker.irio._
